+++ scripts/commands/helm.sh
@@ -167,6 +167,24 @@ helm_wrapper() {
         j=$((j + 1))
     done
 
+    if [ "${EVALUATE_TEMPLATES}" = "true" ]; then
+        set -- "$@" "--post-renderer" "${HELM_BIN}"
+
+        if [ "${HELM_DEBUG:-}" = "1" ] || [ "${HELM_DEBUG:-}" = "true" ] || [ -n "${HELM_SECRETS_DEBUG+x}" ]; then
+            set -- "$@" "--post-renderer-args" "--debug"
+        fi
+
+        set -- "$@" "--post-renderer-args" "secrets"
+        set -- "$@" "--post-renderer-args" "--backend" "--post-renderer-args" "${SECRET_BACKEND}"
+        if [ "${SECRET_BACKEND_ARGS}" != "" ]; then
+            set -- "$@" "--post-renderer-args" "--backend-args" "--post-renderer-args" "${SECRET_BACKEND_ARGS}"
+        fi
+        if [ "${EVALUATE_TEMPLATES_DECODE_SECRETS}" != "" ]; then
+            set -- "$@" "--post-renderer-args" "--evaluate-templates-decode-secrets"
+        fi
+        set -- "$@" "--post-renderer-args" "post-renderer"
+    fi
+
     "${HELM_BIN}" ${TILLER_HOST:+--host "$TILLER_HOST"} "$@"
 }