+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `--ignore-missing-values` Support for evaluating secret references (`vals` backend) in helm templates (requires helm 3.9.0)
+
 ### Fixed
 - Performance issues with large value files (vals backend)